+++ activerecord/test/cases/adapters/postgresql/connection_test.rb
@@ -154,5 +154,46 @@ module ActiveRecord
       end
     end
 
+    def test_set_session_variable_true
+      run_without_connection do |orig_connection|
+        ActiveRecord::Base.establish_connection(orig_connection.deep_merge({:variables => {:debug_print_plan => true}}))
+        set_true = ActiveRecord::Base.connection.exec_query "SHOW DEBUG_PRINT_PLAN"
+        assert_equal set_true.rows, [["on"]]
+      end
+    end
+
+    def test_set_session_variable_false
+      run_without_connection do |orig_connection|
+        ActiveRecord::Base.establish_connection(orig_connection.deep_merge({:variables => {:debug_print_plan => false}}))
+        set_false = ActiveRecord::Base.connection.exec_query "SHOW DEBUG_PRINT_PLAN"
+        assert_equal set_false.rows, [["off"]]
+      end
+    end
+
+    def test_set_session_variable_nil
+      run_without_connection do |orig_connection|
+        # This should be a no-op that does not raise an error
+        ActiveRecord::Base.establish_connection(orig_connection.deep_merge({:variables => {:debug_print_plan => nil}}))
+      end
+    end
+
+    def test_set_session_variable_default
+      run_without_connection do |orig_connection|
+        # This should execute a query that does not raise an error
+        ActiveRecord::Base.establish_connection(orig_connection.deep_merge({:variables => {:debug_print_plan => :default}}))
+      end
+    end
+
+    private
+
+    def run_without_connection
+      original_connection = ActiveRecord::Base.remove_connection
+      begin
+        yield original_connection
+      ensure
+        ActiveRecord::Base.establish_connection(original_connection)
+      end
+    end
+
   end
 end